feat(mapgen-studio): footer consoles — studio (centered) vs live-game (right-docked) with state bridge
The footer separates ownership domains: a single centered Studio console
(status · last run · seed · reroll · auto-run · Run) merges the two
former studio bars; a named, modular GameConsole unit docks right under
a Civ7 eyebrow carrying the live runtime chip + apply-suggestion bridge,
autoplay, Run in Game with status/retry/diagnostics, and the save-deploy
chip — a designated home for future live-game controls. Equal flex side
zones keep the studio console exactly centered while space allows
(verified 860==860 @1720px) and yield left without overlap when the game
console needs room (verified @1199px). All gating/aria/diagnostics and
relation cues (stale ring, Current/Stale/Previous chip, Modified ring)
preserved verbatim; 146 tests + tsc green.
